Finished lab2 on EWS

+++ lab2/num_comps.c
@@ -2,15 +2,15 @@
  
 int main()
 {
-    int val1 = -22;
-    int val2 = 33;
-    int val3 = 12;
-    int val4 = 10;
-    int val5 = 3;
+    double val1 = -22;
+    double val2 = 33;
+    double val3 = 12;
+    double val4 = 10;
+    double val5 = 3;
 
-    int res;
-    int res2;
-    int res3;
+    double res;
+    double res2;
+    double res3;
  
     res = val1 * val2;
     res2 = val3 / val5;
@@ -20,9 +20,9 @@ int main()
     int res5 = (-11) % 3;
 
  
-    printf("%d\n", res);
-    printf("%d\n", res2);
-    printf("%d\n", res3);
+    printf("%f\n", res);
+    printf("%f\n", res2);
+    printf("%f\n", res3);
     printf("%d\n", res4);
     printf("%d\n", res5);
